* Cache bug fixed * Auth bug fixed * Minor cosmetic fix
