+++ app.js
@@ -2,7 +2,7 @@ const express = require('express')
 const app = express()
 const router = require('./routes/index')
 const bodyParser = require('body-parser')
-const { handleCachedData } = require('./middleware/cacheHandler')
+const { handleCachedData } = require('./middleware/cache')
 const path = require('path')
 const config = require('config')
 const { authenticateJwt } = require('./middleware/jwtAuthentication')
