+++ package.json
@@ -32,6 +32,7 @@
     "mysql": "^2.18.1",
     "nock": "^13.1.3",
     "nodemon": "^2.0.13",
+    "path-to-regexp": "^6.2.0",
     "prettier": "^2.5.0",
     "prettier-config": "^1.0.0",
     "redis": "^4.0.0-rc.3",
