+++ services/authentication.js
@@ -1,5 +1,14 @@
 const jwt = require('jsonwebtoken')
 const config = require('config')
+const { pathToRegexp } = require('path-to-regexp')
+
+const isProtectedPath = (reqMethod, reqPath) => {
+  return config.auth.protectedUrls.find((protectedUrl) => {
+    const regExpression = pathToRegexp(protectedUrl.path)
+    const isProtectedUrl = regExpression.exec(reqPath)
+    return isProtectedUrl && (reqMethod === protectedUrl.method)
+  })
+}
 
 const createNewJwt = (username) => {
   return jwt.sign({ username: username }, config.auth.secret, {
@@ -15,4 +24,4 @@ const verifyJwt = (authorization, secret) => {
     })
   })
 }
-module.exports = { createNewJwt, verifyJwt }
+module.exports = { createNewJwt, verifyJwt, isProtectedPath }
